// webapp/resources/sap/esh/search/ui/controls/facets/FacetTypeUI.d.ts
declare module "sap/esh/search/ui/controls/facets/FacetTypeUI" {
    /*!
     * SAPUI5
     * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
     *
     */
    enum FacetTypeUI {
        DataSource = "DataSource",
        Hierarchy = "Hierarchy",
        Attribute = "Attribute",
        QuickSelectDataSource = "QuickSelectDataSource",
        HierarchyStatic = "HierarchyStatic"
    }
}
//# sourceMappingURL=FacetTypeUI.d.ts.map